[turtleNeck-CBJ] 1:1문의 답변 작성 백앤드 detail : 1:1문의 답변 작성 백앤드

+++ jbell_backend/src/main/java/jbell/qna/dto/QnaAnswerCreate.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class QnaAnswerCreate {
 
-	private int qnaId;					// 연결된 문의글의 식별자 (FK)
+	private Long qnaId;					// 연결된 문의글의 식별자 (FK)
 	private String content;				// 답변 본문 내용
 	private String userId;				// 답변을 작성한 관리자 ID (FK)
 }
