+++ jbell_backend/src/main/java/jbell/qna/dto/QnaAnswerUpdate.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class QnaAnswerUpdate {
 
-	private int qnaAnswerId;    // 답변 고유 시별자 (보통 path variable로도 가능)
+	private Long qnaAnswerId;    // 답변 고유 시별자 (보통 path variable로도 가능)
 	private String content;		// 수정할 답변 본문 내용
 }
