+++ jbell_backend/src/main/java/jbell/qna/mapper/QnaMapper.java
@@ -28,13 +28,13 @@ public interface QnaMapper {
      * QnA 상세 조회
      * 매핑 XML ID: getQnaDetail
      */
-    QnaDetail getQnaDetail(@Param("qnaId") int qnaId);
+    QnaDetail getQnaDetail(@Param("qnaId") Long qnaId);
     
     /**
      * QnA 수정
      * @param qnaId 대상 문의 ID
      * @param update 수정할 데이터 객체
      */
-    void updateQna(@Param("qnaId") int qnaId, @Param("update") jbell.qna.dto.QnaUpdate update);
+    void updateQna(@Param("qnaId") Long qnaId, @Param("update") jbell.qna.dto.QnaUpdate update);
 
 }
